+++ src/main/java/babysnoozer/events/SetSnoozingEndPosEvent.java
@@ -4,13 +4,13 @@ package babysnoozer.events;
  * Created by Alexander Bischof on 12.01.15.
  */
 public class SetSnoozingEndPosEvent {
-  private final int endPos;
+  private final long endPos;
 
-  public SetSnoozingEndPosEvent(int endPos) {
+  public SetSnoozingEndPosEvent(long endPos) {
 	this.endPos = endPos;
   }
 
-  public int getEndPos() {
+  public long getEndPos() {
 	return endPos;
   }
 }
