+++ src/main/java/babysnoozer/handlers/RopeDistance.java
@@ -4,33 +4,33 @@ package babysnoozer.handlers;
  * Created by Alexander Bischof on 30.01.15.
  */
 public class RopeDistance {
-    private int startPos;
-    private int endPos;
+    private long startPos;
+    private long endPos;
 
-    public int getStartPos() {
+    public long getStartPos() {
         return startPos;
     }
 
-    public void setStartPos(int startPos) {
+    public void setStartPos(long startPos) {
         this.startPos = startPos;
     }
 
-    public int getEndPos() {
+    public long getEndPos() {
         return endPos;
     }
 
-    public void setEndPos(int endPos) {
+    public void setEndPos(long endPos) {
         this.endPos = endPos;
     }
 
-    public int getPosBy(int percentage){
+    public long getPosBy(int percentage){
         if (percentage < 0 || percentage > 100){
             throw new IllegalArgumentException("percentage between 0 and 100 expected");
         }
 
-        int distance = Math.abs(endPos-startPos);
+        long distance = Math.abs(endPos-startPos);
         distance = percentage * distance /100;
 
-        return (int) (Math.min(endPos, startPos) + distance);
+        return (Math.min(endPos, startPos) + distance);
     }
 }
